Engine fix

+++ src/ai/nodes/IsEnemyVisible.hpp
@@ -25,7 +25,7 @@ class IsEnemyVisible : public BT::SyncActionNode {
 		return {
 			BT::InputPort<Entity>("entity"),
 			BT::OutputPort<Entity>("otherEntity"),
-			BT::OutputPort<Vec2d>("otherPosition"),
+			BT::OutputPort<Vec2i>("otherPosition"),
 			BT::OutputPort<Rotation>("direction")
 		};
 		// clang-format on
@@ -40,16 +40,16 @@ class IsEnemyVisible : public BT::SyncActionNode {
 			return BT::NodeStatus::FAILURE;
 		}
 
-		const Vec2d &position = ecs.getComponent<Positionable>(entity).position;
+		const Vec2i &position = ecs.getComponent<Positionable>(entity).position;
 		const Rotation &rotation = ecs.getComponent<Rotatable>(entity).rotation;
 		const Entity &otherEntity = vision.visibleEnemies[0];
-		const Vec2d &otherPosition = ecs.getComponent<Positionable>(otherEntity).position;
+		const Vec2i &otherPosition = ecs.getComponent<Positionable>(otherEntity).position;
 
 		// TODO: Is this actually the best place to do this?
 		Rotation direction = calculateDirection(position, otherPosition, rotation);
 
 		setOutput<Entity>("otherEntity", otherEntity);
-		setOutput<Vec2d>("otherPosition", otherPosition);
+		setOutput<Vec2i>("otherPosition", otherPosition);
 		setOutput<Rotation>("direction", direction);
 
 		return BT::NodeStatus::SUCCESS;
@@ -68,7 +68,7 @@ class IsEnemyVisible : public BT::SyncActionNode {
 		return exp.value();
 	}
 
-	Rotation calculateDirection(const Vec2d &start, const Vec2d &end, const Rotation &currentRotation)
+	Rotation calculateDirection(const Vec2i &start, const Vec2i &end, const Rotation &currentRotation)
 	{
 		// Calculate the difference between the points
 		double dx = end.x - start.x;
