+++ src/ui/TextDialog.hpp
@@ -2,7 +2,7 @@
 
 #include "../constants.hpp"
 #include "../engine/Engine.hpp"
-#include "../engine/Vec2d.hpp"
+#include "../engine/Vec2i.hpp"
 #include "../modules/MenuStack.hpp"
 #include "UIElement.hpp"
 #include <string>
@@ -31,7 +31,7 @@ class TextDialog final : public UIElement {
 
 	void render()
 	{
-		renderContainer(Vec2d{x, y}, menuWidth_, menuHeight_);
+		renderContainer(Vec2i{x, y}, menuWidth_, menuHeight_);
 		renderText();
 	}
